requirements.txt
numpy>=1.19.0
pandas>=1.3.0
gymnasium>=0.26.0
stable-baselines3>=1.5.0
torch>=1.10.0
matplotlib>=3.4.0
tensorboard>=2.6.0
Flask>=2.0.0 